Minor changes to avoid compiler warnings.

+++ dlib/image_processing/shape_predictor.h
@@ -711,7 +711,7 @@ namespace dlib
         {
             using namespace impl;
             std::deque<std::pair<unsigned long, unsigned long> > parts;
-            parts.push_back(std::make_pair(0, samples.size()));
+            parts.push_back(std::make_pair(0, (unsigned long)samples.size()));
 
             impl::regression_tree tree;
